added notification

+++ app/Http/Controllers/AdminController.php
@@ -7,6 +7,8 @@ use Dingo\Api\Routing\Helpers;
 use Response;
 use Illuminate\Support\Collection;
 use App\Transformers\UsersTransformer;
+use Illuminate\Support\Facades\Notification;
+use App\Notifications\UserNotification;
 class AdminController extends Controller
 {
     /**
@@ -15,29 +17,37 @@ class AdminController extends Controller
      * @return void
      */
     use Helpers;
-     public function index()
+    public function index()
     {
-        return $this->response->collection(Collection::make(User::where('role','admin')->get()), new UsersTransformer);
+        // $this->authorize('index', User::class);
+
+        return $this->response->collection(Collection::make(User::all()), new UsersTransformer);
     }
     
     public function show($id)
     {
-    
-     return $this->response->collection(Collection::make(User::where('role','admin')->where('id',$id)->get()), new UsersTransformer);
+     // $this->authorize('show', User::class);
+         $email = User::where('id',$id)->first();
+         $email->notify(new UserNotification($email));
+     return $this->response->collection(Collection::make(User::where('id',$id)->get()), new UsersTransformer);
     }
     public function create(Request $request)
     {
+        // $this->authorize('create', User::class);
         $user = new User;
         $user->name= $request->name;
         $user->email = $request->email;
-        $user->password = $request->password;
+        $user->password = app('hash')->make($request->password);
         $user->role = $request->role;
        
         $user->save();
-        return $this->response->collection(Collection::make($user), new UsersTransformer);
+        $email = User::where('email',$user->email)->first();
+         $email->notify(new UserNotification($email));
+        return response()->json($user);
     }
     public function update(Request $request, $id)
     {
+        // $this->authorize('update', User::class);
         $user= user::find($id);
         $user->name= $request->input('name');
         $user->email = $request->input('email');
@@ -45,10 +55,12 @@ class AdminController extends Controller
         $user->role = $request->input('role');
        
         $user->save();
-         return $this->response->collection(Collection::make($user), new UsersTransformer);
+         return response()->json($user);
+         // $this->response->collection(Collection::make($user), new UsersTransformer);
     }
     public function destroy($id)
     {
+        $this->authorize('destroy', User::class);
         $user = User::find($id);
         $user->delete();
         return response()->json('user deleted successfully');
