+++ app/Http/Controllers/UserController.php
@@ -1,9 +1,9 @@
 <?php
 
 namespace App\Http\Controllers;
+use Illuminate\Http\Request;
 use App\User;
 use App\Contact;
-use Illuminate\Http\Request;
 use Dingo\Api\Routing\Helpers;
 use Response;
 use Illuminate\Support\Collection;
@@ -22,22 +22,26 @@ class UserController extends Controller
        
         //
     }
-    public function show($id)
+    // public function show($id)
+    // {
+    //     $this->authorize('create', User::class);
+    //     $user=User::where('id',$id)->get();
+    //      return response()->json($user);
+    // }
+    public function index($id)
     {
-        $user=User::where('id',$id)->get();
-         return response()->json($user);
-    }
-    public function showcontacts($id)
-    {
+        $this->authorize('create', Contact::class);
         $contacts=Contact::where('user_id',$id)->get();
          return $this->response->collection(Collection::make($contacts),new ContactsTransformer);
     }
-     public function showSpecificContact($id,$contactid)
+     public function show($id,$contactid)
     {
+        $this->authorize('show', User::class);
          return $this->response->collection(Collection::make(Contact::where('user_id',$id)->where('id',$contactid)->get()), new ContactsTransformer);
     }
     public function create(Request $request,$id)
     {
+        $this->authorize('create', Contact::class);
         $contact = new Contact;
         $contact->name= $request->name;
         $contact->email = $request->email;
@@ -53,11 +57,14 @@ class UserController extends Controller
        
        
         $contact->save();
-        return $this->response->collection(Collection::make($contact), new ContactsTransformer);
+        $transformedcontact = Contact::where('email',$contact->email)->get();
+
+        return response()->json($transformedcontact);
     }
-    public function update(Request $request, $id)
+    public function update(Request $request, $id,$contactid)
     {
-        $contact= Contact::find($id);
+        $this->authorize('update', Contact::class);
+        $contact= Contact::find($contactid);
         $contact->name= $request->input('name');
         $contact->email = $request->input('email');
         $contact->phone = $request->input('phone');
@@ -66,14 +73,16 @@ class UserController extends Controller
         $contact->title = $request->input('title');
         $contact->notes = $request->input('notes');
         $contact->meetings = $request->input('meetings');
-        $contact->user_id = $request->$id;
+        
         $contact->opportunities = $request->input('opportunities');
         $contact->prefered_notification_method = $request->input('prefered_notification_method');
         $contact->save();
-        return $this->response->collection(Collection::make($contact), new ContactsTransformer);
+        $transformedcontact = Contact::where('id',$contactid)->get();
+        return $this->response->collection(Collection::make($transformedcontact), new ContactsTransformer);
     }
     public function destroy($id,$contactid)
     {
+        $this->authorize('destroy', Contact::class);
         $contact = Contact::where('user_id',$id)->find($id);
         $contact->delete();
